Be more strict with requirements: scipy, pandas (#318)
* Be more strict with requirements: scipy, pandas

Closes #317

* Fix

* Fix

* moar fix

+++ tests/test_core.py
@@ -876,8 +876,8 @@ def test_verify_complex_cohorts(chunksize: int) -> None:
     chunk_cohorts = find_group_cohorts(by - 1, (chunks,))
     chunks_ = np.sort(np.concatenate(tuple(chunk_cohorts.keys())))
     groups = np.sort(np.concatenate(tuple(chunk_cohorts.values())))
-    assert_equal(np.unique(chunks_), np.arange(len(chunks), dtype=int))
-    assert_equal(groups, np.arange(366, dtype=int))
+    assert_equal(np.unique(chunks_).astype(np.int64), np.arange(len(chunks), dtype=np.int64))
+    assert_equal(groups.astype(np.int64), np.arange(366, dtype=np.int64))
 
 
 @requires_dask
